Refactor pkg/generate
